+++ Pandora/Paradigm/Primary/Transformer/Yoneda.hs
@@ -39,7 +39,7 @@ instance Extractable t (->) => Extractable (Yoneda t) (->) where
 	extract (Yoneda f) = extract $ f identity
 
 instance Liftable Yoneda where
-	lift x = Yoneda (<$> x)
+	lift x = Yoneda (-<$>- x)
 
 instance (Extractable t (->), Pointable t (->), Extractable u (->) , Pointable u (->)) => Adjoint (Yoneda t) (Yoneda u) where
 	x -| f = point . f . point # x